fix test

+++ internal/core/src/index/TextMatchIndex.cpp
@@ -299,7 +299,7 @@ TextMatchIndex::MatchQuery(const std::string& query) {
         Reload();
     }
 
-    TargetBitmap bitset{Count()};
+    TargetBitmap bitset{static_cast<size_t>(Count())};
     // The count opeartion of tantivy may be get older cnt if the index is committed with new tantivy segment.
     // So we cannot use the count operation to get the total count for bitmap.
     // Just use the maximum offset of hits to get the total count for bitmap here.
@@ -314,7 +314,7 @@ TextMatchIndex::PhraseMatchQuery(const std::string& query, uint32_t slop) {
         Reload();
     }
 
-    TargetBitmap bitset{Count()};
+    TargetBitmap bitset{static_cast<size_t>(Count())};
     // The count opeartion of tantivy may be get older cnt if the index is committed with new tantivy segment.
     // So we cannot use the count operation to get the total count for bitmap.
     // Just use the maximum offset of hits to get the total count for bitmap here.
